requirements.txt
requests
python-dotenv
pandas
duckdb
